Rename inspect.py module to fix name clash
